refactor de componentes
